requirements.txt
appnope==0.1.0
backcall==0.2.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.3.1
matplotlib==3.4.1
numexpr==2.7.1
numpy==1.19.4
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.4.1
-e git+https://github.com/shahcompbio/scgenome.git@687627ae1b15f35a415ddc550d2ea686d2f1582f#egg=scgenome
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
tables==3.6.1
traitlets==5.0.5
wcwidth==0.2.5
-e git+https://github.com/amcpherson/wgs_analysis.git@e9c35fa47334fa431b010ebe5744c7bfada78b73#egg=wgs_analysis
ya==0.0.2
